pyproject.toml
[project]
name = "project-specify-cli"
version = "0.0.1"
description = "Project planning toolkit with spec-driven development for all AI IDEs. Forked from GitHub Spec Kit."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
keywords = [
    "ai",
    "spec-driven-development",
    "project-planning",
    "claude",
    "cursor",
    "copilot",
    "specification",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Code Generators",
]
# These match the original spec-kit dependencies
dependencies = [
    "typer",                    # CLI framework (original uses this)
    "rich",                     # Beautiful terminal output
    "httpx[socks]",             # HTTP client for fetching templates
    "platformdirs",             # Cross-platform directories
    "readchar",                 # Character input handling
    "truststore>=0.10.4",       # SSL/TLS certificate handling
    "pyyaml>=6.0",              # For monorepo workspace parsing and MCP discovery
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",      # Mock and spy support for pytest
    "responses>=0.24.0",         # HTTP request mocking
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
# CHANGED: New entry point name
project-specify = "specify_cli.main:main"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/project-specify"
Documentation = "https://github.com/YOUR_USERNAME/project-specify#readme"
Repository = "https://github.com/YOUR_USERNAME/project-specify.git"
Issues = "https://github.com/YOUR_USERNAME/project-specify/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specify_cli"]

# CRITICAL: Include non-Python files in the package
# Without this, agents/*.md, templates/*.md, scripts/*.sh won't be included
[tool.hatch.build.targets.wheel.force-include]
"src/specify_cli/agents" = "specify_cli/agents"
"src/specify_cli/templates" = "specify_cli/templates"
"src/specify_cli/scripts" = "specify_cli/scripts"

# Belt-and-suspenders: Also use include patterns
[tool.hatch.build]
include = [
    "src/specify_cli/**/*.py",
    "src/specify_cli/agents/**/*.md",
    "src/specify_cli/templates/**/*.md",
    "src/specify_cli/scripts/**/*.sh",
    "src/specify_cli/scripts/**/*.ps1",
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "**/.DS_Store",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
